Restructured Folders
The Models folder was getting out of hand, so I made subfolders. Figured now was a good time to do this since I'm going to be fighting with CircleCI next time I work on this I'm sure.
