+++ SharpenUp.Tests/MonitorsTests.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using SharpenUp.Client;
 using SharpenUp.Common;
-using SharpenUp.Common.Models;
+using SharpenUp.Common.Models.Monitors;
 using SharpenUp.Common.Types;
 using Xunit;
