+++ SharpenUp/Common/IUptimeManager.cs
@@ -1,5 +1,7 @@
 ﻿using System.Threading.Tasks;
-using SharpenUp.Common.Models;
+using SharpenUp.Common.Models.Accounts;
+using SharpenUp.Common.Models.Alerts;
+using SharpenUp.Common.Models.Monitors;
 
 namespace SharpenUp.Common
 {
