+++ SharpenUp/Common/Models/Log.cs
@@ -1,6 +1,5 @@
-﻿using System;
+﻿using Newtonsoft.Json;
 using SharpenUp.Common.Types;
-using Newtonsoft.Json;
 
 namespace SharpenUp.Common.Models
 {
