+++ SharpenUp/Common/Models/Monitors/MonitorsRequest.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using SharpenUp.Common.Types;
 
-namespace SharpenUp.Common.Models
+namespace SharpenUp.Common.Models.Monitors
 {
     public class MonitorsRequest
     {
