+++ SharpenUp/Common/Models/Monitors/MonitorsResult.cs
@@ -2,7 +2,7 @@
 using Newtonsoft.Json;
 using SharpenUp.Common.Types;
 
-namespace SharpenUp.Common.Models
+namespace SharpenUp.Common.Models.Monitors
 {
     public class MonitorsResult
     {
